third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristic/characteristic-found.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return getHealthThermometerDevice()
    .then(({device}) => device.gatt.getPrimaryService('health_thermometer'))
    .then(service => {
      return Promise.all([
        service.getCharacteristic(measurement_interval.alias),
        service.getCharacteristic(measurement_interval.name),
        service.getCharacteristic(measurement_interval.uuid)])
        .then(characteristics => {
          characteristics.forEach(characteristic => {
            assert_equals(
              characteristic.uuid, measurement_interval.uuid,
              'Characteristic UUID should be the same as requested UUID.');
            assert_equals(
              characteristic.service, service,
              'Characteristic service should be the same as service.');
          });
       });
    });
}, 'Request for characteristic. Should return right characteristic');
</script>
